More compile fixes.
