+++ src/makefile.linux-mingw
@@ -199,33 +199,33 @@ obj/scrypt-x86_64.o: scrypt-x86_64.S
 obj/scrypt-arm.o: scrypt-arm.S
 	$(CXX) -c $(xCXXFLAGS) -MMD -o $@ $<
 clean:
-	-rm -f darksilkd
-	-rm -f obj/*.o
-	-rm -f obj/*.P
-	-rm -f obj/*.d
-	-rm -f obj/build.h
-	-rm -f obj/crypto/*.o
-	-rm -f obj/crypto/*.P
-	-rm -f obj/crypto/*.d
-	-rm -f obj/primitives/*.o
-	-rm -f obj/primitives/*.P
-	-rm -f obj/primitives/*.d
-        -rm -f obj/consensus/*.o
-        -rm -f obj/consensus/*.P
-        -rm -f obj/consensus/*.d
-        -rm -f obj/scripts/*.o
-        -rm -f obj/scripts/*.P
-        -rm -f obj/scripts/*.d
-        -rm -f obj/wallet/*.o
-        -rm -f obj/wallet/*.P
-        -rm -f obj/wallet/*.d
-	-rm -f obj/crypto/argon2/*.o
-	-rm -f obj/crypto/argon2/*.P
-	-rm -f obj/crypto/argon2/*.d
-	-rm -f obj/crypto/argon2/blake2/*.o
-	-rm -f obj/crypto/argon2/blake2/*.P
-	-rm -f obj/crypto/argon2/blake2/*.d
-	cd leveldb && TARGET_OS=OS_WINDOWS_CROSSCOMPILE $(MAKE) clean && cd ..
-	cd secp256k1 && TARGET_OS=OS_WINDOWS_CROSSCOMPILE $(MAKE) clean && cd ..
+-rm -f darksilkd
+-rm -f obj/*.o
+-rm -f obj/*.P
+-rm -f obj/*.d
+-rm -f obj/build.h
+-rm -f obj/crypto/*.o
+-rm -f obj/crypto/*.P
+-rm -f obj/crypto/*.d
+-rm -f obj/primitives/*.o
+-rm -f obj/primitives/*.P
+-rm -f obj/primitives/*.d
+-rm -f obj/consensus/*.o
+-rm -f obj/consensus/*.P
+-rm -f obj/consensus/*.d
+-rm -f obj/scripts/*.o
+-rm -f obj/scripts/*.P
+-rm -f obj/scripts/*.d
+-rm -f obj/wallet/*.o
+-rm -f obj/wallet/*.P
+-rm -f obj/wallet/*.d
+-rm -f obj/crypto/argon2/*.o
+-rm -f obj/crypto/argon2/*.P
+-rm -f obj/crypto/argon2/*.d
+-rm -f obj/crypto/argon2/blake2/*.o
+-rm -f obj/crypto/argon2/blake2/*.P
+-rm -f obj/crypto/argon2/blake2/*.d
+cd leveldb && TARGET_OS=OS_WINDOWS_CROSSCOMPILE $(MAKE) clean && cd ..
+cd secp256k1 && TARGET_OS=OS_WINDOWS_CROSSCOMPILE $(MAKE) clean && cd ..
 
 FORCE:
